+++ src/unraid_api/__init__.py
@@ -88,4 +88,4 @@ __all__ = [
     "VmDomain",
 ]
 
-__version__ = "1.3.0"
+__version__ = "1.3.1"